requirements.txt
python-pptx>=0.6.21
pandas>=1.5.0
Pillow>=9.0.0
